Update scratch_9.py
hello,
this change is small and I mad this contribution to just a motivation to you how the open source is work in GitHub and inter net.

its really a good project but still there is a small issues. this project will need some sort of storing mechanism. while we make shot version of url it need to store the key value pairs permanently so that user can get the expand version after he need view or  redirect to it . with out this whatever the output , its hard to believe that the output is url until review of code.
this can be solved  using  files or data base connection in code 

all the best.